Correction of issue caused by 12bit mask
In our lab, we recently found a set of MR series that were 12bit in DICOM with negative values. At first, these negative values appeared to be treated as unsigned short with wrap around (values in the 4095 vicinity). Review of the source code revealed that the idea of discarding potentially leaked information in the upper 4 bits also removed the proper sign and magnitude of the intensity value due to the truncation by the mask. Since the nifti header lacks a defined way to let other tools know that the data is in the 12bit range (instead of the signed 16bit range), I propose to use bit shifting to accomplish the same task as the mask while allowing for preservation of the sign information in the upper 4 bits of the 16 bit range. Issue was a direct result of https://github.com/rordenlab/dcm2niix/issues/251 and the fact the data is saved in the 16bit range.