Making previous change optimization resistant
Per Compiler Explorer (https://godbolt.org/), older versions of gcc and clang might optimize the combined bit shift operation out. Meaning, it would be the same as if we did not attempt to truncate the upper 4 bits. It is not clear if that was an aberration of testing with CE, but the risk is enough for me to rearrange the code such that we remain faithful to the initial purpose of the function. Do note that newer compilers will optimize these shifts to use SIMD, but they will preserve the shift left and shift right instructions. Meaning, newer compilers might give us better performance here without sacrificing the correctness of the code. Tested with GCC, CLANG, and MSVC for x86_64 using -O2 optimization flag.