This Fix corrects the TotalReadoutTime and EffectiveEchoSpacing fields in the BIDS (.json) sidecar in cose of image interpolation.
TotalReadoutTime and EffectiveEchoSpacing were correct when there was
no interpolation (i.e. AcquisitionMatrixPE = ReconMatrixPE).
However, when interpolation was used, this values were incorrect.
When interpolation was used, the estimated fieldmap given by TOPUP
was ReconMatrixPE/AcquisitionMatrixPE times a measured fieldmap.
Also, the same interpolated images corrected with a B0Map based method
via FSL's epi_reg were clearly under-corrected.
It turned out that the problem was comming from TotalReadoutTime and
EffectiveEchoSpacing which were not taking ReconMatrixPE correctly
into account.
This is a Fix for that.

GE/README.md was modified to clarified how TotalReadoutTime and
EffectiveEchoSpacing are now computed.
Some intermediate variable were also introduced and this described.
This should help he final user to better understand.

console/nii_dicom_batch.cpp was modified to implement the changes on
TotalReadoutTime and EffectiveEchoSpacing.
The intermediate variables described in README were also added into
the json BIDS sidecar (i.e. EchoSpacingMicroSecondsGE,
NotPhysicalNumberOfAcquiredPELinesGE, NotPhysicalTotalReadOutTimeGE).
As the TotalReadoutTime definition is not intuitive this should
help the user to understand what is going on.

On branch dev-fix-TotalReadoutTimeGE-with-interpolation
Changes to be committed:
	modified:   GE/README.md
	modified:   console/nii_dicom_batch.cpp